+++ Runtime/Subsystems/Anchor/ARAnchorFeature.cs
@@ -16,7 +16,7 @@ namespace UnityEngine.XR.OpenXR.Features.Meta
         Company = Constants.k_CompanyName,
         Desc = "AR Foundation anchor support on Meta Quest devices",
         DocumentationLink = Constants.DocsUrls.k_AnchorsUrl,
-        OpenxrExtensionStrings = Constants.OpenXRExtensions.k_XR_FB_spatial_entity,
+        OpenxrExtensionStrings = k_OpenXRRequestedExtensions,
         Category = FeatureCategory.Feature,
         FeatureId = featureId,
         Version = "0.1.0")]
@@ -27,6 +27,16 @@ namespace UnityEngine.XR.OpenXR.Features.Meta
         /// The feature id string. This is used to give the feature a well known id for reference.
         /// </summary>
         public const string featureId = "com.unity.openxr.feature.arfoundation-meta-anchor";
+
+        /// <summary>
+        /// The set of OpenXR spec extension strings to enable, separated by spaces.
+        /// For more information, refer to
+        /// <see href="https://docs.unity3d.com/Packages/com.unity.xr.openxr@1.6/manual/features.html#enabling-openxr-spec-extension-strings"/>.
+        /// </summary>
+        /// <seeaslo href="https://developer.oculus.com/documentation/native/android/mobile-scene-api-ref/"/>
+        const string k_OpenXRRequestedExtensions =
+            Constants.OpenXRExtensions.k_XR_FB_spatial_entity + " " +
+            Constants.OpenXRExtensions.k_XR_FB_spatial_entity_storage;
 
         static List<XRAnchorSubsystemDescriptor> s_AnchorDescriptors = new();