+++ Tests/CodeSamples/AnchorsSamples.cs
@@ -1,11 +1,12 @@
 using UnityEngine.TestTools;
 using UnityEngine.XR.ARFoundation;
+using UnityEngine.XR.OpenXR.NativeTypes;
 
 namespace UnityEngine.XR.OpenXR.Features.Meta.Tests
 {
     public class AnchorsSamples
     {
-        public async void AddAnchorAsync_ConvertResultToXrResult()
+        public async void AddAnchorAsync_GetNativeXRResult()
         {
 #region anchors_resultstatus_xrresult
             // This is not optimal. For better performance, reuse a saved reference instead.
@@ -16,8 +17,8 @@ namespace UnityEngine.XR.OpenXR.Features.Meta.Tests
             var pose = new Pose(Vector3.zero, Quaternion.identity);
             var result = await anchorManager.TryAddAnchorAsync(pose);
 
-            // To access OpenXR error and success codes, convert status to XrResult
-            var xrResult = result.status.AsXrResult();
+            // To access OpenXR error and success codes, use the nativeStatusCode property
+            var xrResult = (XrResult)result.status.nativeStatusCode;
 
             // Prints "Success", or the error or success code associated with this operation
             Debug.Log(xrResult);